feat: enhance voting functionality with new features and test coverage
- Added support for vote staggering in k6 cloud load tests (new input `vote_stagger_seconds`).
- Enabled separate channels for public and judge voting in voting sessions.
- Introduced many-to-many associations between participants and voter accounts.
- Improved frontend logic to manage voting states for public and judge channels.
- Updated database seed and results APIs to handle team voting scenarios.
- Enhanced test suite with comprehensive coverage for new features.
- Fixed frontend URLs to use `eventId` instead of `competitionId`.

+++ backend/app/schemas/results.py
@@ -11,7 +11,7 @@ class ResultEntryRead(BaseModel):
     final_score: float
     public_score: float
     judge_score: float
-    public_votes: int
+    public_votes: float
     judge_votes_count: int
     details: dict[str, Any]
